move rtree and canvas_spy sources to Fuchsia sub-directory
