+++ ci/licenses_golden/excluded_files
@@ -74,7 +74,6 @@
 ../../../flutter/flow/layers/transform_layer_unittests.cc
 ../../../flutter/flow/mutators_stack_unittests.cc
 ../../../flutter/flow/raster_cache_unittests.cc
-../../../flutter/flow/rtree_unittests.cc
 ../../../flutter/flow/skia_gpu_object_unittests.cc
 ../../../flutter/flow/surface_frame_unittests.cc
 ../../../flutter/flow/testing
@@ -212,7 +211,6 @@
 ../../../flutter/runtime/no_dart_plugin_registrant_unittests.cc
 ../../../flutter/runtime/type_conversions_unittests.cc
 ../../../flutter/shell/common/animator_unittests.cc
-../../../flutter/shell/common/canvas_spy_unittests.cc
 ../../../flutter/shell/common/context_options_unittests.cc
 ../../../flutter/shell/common/dl_op_spy_unittests.cc
 ../../../flutter/shell/common/engine_unittests.cc
@@ -292,6 +290,7 @@
 ../../../flutter/shell/platform/fuchsia/flutter/accessibility_bridge_unittest.cc
 ../../../flutter/shell/platform/fuchsia/flutter/build/asset_package.py
 ../../../flutter/shell/platform/fuchsia/flutter/build/gen_debug_wrapper_main.py
+../../../flutter/shell/platform/fuchsia/flutter/canvas_spy_unittests.cc
 ../../../flutter/shell/platform/fuchsia/flutter/component_v2_unittest.cc
 ../../../flutter/shell/platform/fuchsia/flutter/focus_delegate_unittests.cc
 ../../../flutter/shell/platform/fuchsia/flutter/fuchsia_intl_unittest.cc
@@ -303,6 +302,7 @@
 ../../../flutter/shell/platform/fuchsia/flutter/platform_view_unittest.cc
 ../../../flutter/shell/platform/fuchsia/flutter/pointer_delegate_unittests.cc
 ../../../flutter/shell/platform/fuchsia/flutter/pointer_injector_delegate_unittest.cc
+../../../flutter/shell/platform/fuchsia/flutter/rtree_unittests.cc
 ../../../flutter/shell/platform/fuchsia/flutter/runner_tzdata_missing_unittest.cc
 ../../../flutter/shell/platform/fuchsia/flutter/runner_tzdata_unittest.cc
 ../../../flutter/shell/platform/fuchsia/flutter/tests
