+++ flow/BUILD.gn
@@ -76,8 +76,6 @@ source_set("flow") {
     "raster_cache_key.h",
     "raster_cache_util.cc",
     "raster_cache_util.h",
-    "rtree.cc",
-    "rtree.h",
     "skia_gpu_object.h",
     "surface.cc",
     "surface.h",
@@ -171,7 +169,6 @@ if (enable_unittests) {
       "layers/transform_layer_unittests.cc",
       "mutators_stack_unittests.cc",
       "raster_cache_unittests.cc",
-      "rtree_unittests.cc",
       "skia_gpu_object_unittests.cc",
       "surface_frame_unittests.cc",
       "testing/mock_layer_unittests.cc",
