+++ shell/platform/darwin/ios/framework/Source/FlutterPlatformViews.mm
@@ -10,7 +10,6 @@
 #include <string>
 
 #include "flutter/common/graphics/persistent_cache.h"
-#include "flutter/flow/rtree.h"
 #include "flutter/fml/platform/darwin/scoped_nsobject.h"
 #import "flutter/shell/platform/darwin/common/framework/Headers/FlutterChannels.h"
 #import "flutter/shell/platform/darwin/ios/framework/Source/FlutterOverlayView.h"
