+++ shell/platform/fuchsia/flutter/BUILD.gn
@@ -58,6 +58,8 @@ template("runner_sources") {
     sources = [
       "accessibility_bridge.cc",
       "accessibility_bridge.h",
+      "canvas_spy.cc",
+      "canvas_spy.h",
       "component_v2.cc",
       "component_v2.h",
       "engine.cc",
@@ -94,6 +96,8 @@ template("runner_sources") {
       "pointer_injector_delegate.cc",
       "pointer_injector_delegate.h",
       "program_metadata.h",
+      "rtree.cc",
+      "rtree.h",
       "runner.cc",
       "runner.h",
       "software_surface.cc",
@@ -499,6 +503,7 @@ if (enable_unittests) {
 
     sources = [
       "accessibility_bridge_unittest.cc",
+      "canvas_spy_unittests.cc",
       "component_v2_unittest.cc",
       "flutter_runner_fakes.h",
       "focus_delegate_unittests.cc",
@@ -507,6 +512,7 @@ if (enable_unittests) {
       "platform_view_unittest.cc",
       "pointer_delegate_unittests.cc",
       "pointer_injector_delegate_unittest.cc",
+      "rtree_unittests.cc",
       "tests/engine_unittests.cc",
       "tests/fake_flatland_unittests.cc",
       "tests/fake_session_unittests.cc",
